fix(core): inconsistent analytics string across operating systems
The gzip header includes an “OS” byte that indicates which operating system was used to perform the operation. This means that the CDK Metadata analytics string would be different for the same CDK app in two different operating systems.

To address this, we explicitly set the OS flag to 255 (unknown).

Fixes #15322