chore(resolve-version): export the alphaVersion (#16321)
This will be used by a change in `tools/individual-pkg-gen/copy-files-removing-deps.ts` on the `v2-main` branch, so that the `transform` step can correctly set the `version` key and the version for each dependency in each alpha module's `package.json` file.

See https://github.com/aws/aws-cdk/pull/16322 for usage. 

Part of #15591 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*