fix(core): asset hash of symlinked dir is wrong
When the root directory of an asset is a symlink (such as can happen in
CDK Pipelines), the asset hash calculation incorrectly doesn't follow
the symlink and hashes the link itself it instead.

This leads to the asset hash never changing, even though the files
inside the directory do change.

Instead, we resolve the asset root dir, and make sure to hash the target
directory on disk. Handling of symlinks found *inside* the target dir
remains unchanged.