feat(ecs-service-extensions): Publish Extension (#16326)
----
This PR adds a new service extension, `PublisherExtension`. This extension can be added to a service to allow it to publish events to SNS Topics. (This PR when paired with #16049 can be used to set up the pub/ sub architecture pattern)

It sets up publish permissions for the service to be able to publish events to the topics provided. The user can also provide a list of accounts that will be given permissions to subscribe to the given topics.

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*