chore(aws-cdk): backport %construct-versions% for init templates
When the constructs compatibility layer was removed on the v2 branch (#12054), a
change was made to the init templates to support a flexible constructs version
(either ^3 or ^10). These changes were never back-ported to v1, leading to the
situation where there are (unnecessary) differences on the v2 templates between
the v1 and v2 branches.

Backported these changes manually (for package.json and init.ts), and then by
diffing the init templates directory between master and v2-main.