chore: migrate all remaining modules from NodeUnit to Jest (#16585)
Migrates `aws-cloudformation`, `cfnspec`, and `aws-codebuild` to jest.

In `aws-cloudformation`, jest does not like the idea of nested tests; however, as the scope of this PR is to migrate, I have added `eslint-ignore` to the relevant lines. The linter error in question is [valid-describe](https://github.com/jest-community/eslint-plugin-jest/blob/main/docs/rules/valid-describe.md). 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*