fix(config): add SourceArn condition to Lambda permission
According to
[AWS Config best practices](https://docs.aws.amazon.com/config/latest/developerguide/evaluate-config_develop-rules_nodejs.html#restricted-lambda-policy),
we should add a `SourceAccount` condition to the Lambda Permission we create in `CustomRule`.

Note that we cannot add the `SourceArn` condition,
because that would cause a cyclic dependency between the `LambdaPermission` resource,
and the `Rule` resource
(as the `Rule` can only be created _after_ the `LambdaPermission` has been created -
this is validated by the AWS Config service).