chore: enable publishing of independent alpha modules (#16674)
This is the (hopefully!) final commit to enable us to independently publish the
experimental and dev-preview (collectively "alpha") modules as independent
packages.

After this change, the next automated v2 release will publish both aws-cdk-lib,
as well as one `@aws-cdk/aws-foobar-alpha` module for each service that is
either in experimental or dev-preview.