feat(aws-ec2): userdata cfn-signal signal resource which is different than the attached resource (#16264)
When using `cfn-init` in EC2 UserData with a LaunchTemplate and an AutoScalingGroup the attached resource (used with `cfn-init`) is the LaunchTemplate but the signal (`cfn-signal`) should call the AutoScalingGroup.

`CloudFormationInit::attach` method supports only a single resource for both init and signal - blocking the usage of LaunchTemplate and AutoScalingGroup scenario.
The commit avoids a breaking change in the function signature, by adding an optional `signalResource` to the `AttachInitOptions` parameters argument


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*