chore: (revert) chore: move modules to assertions (#16671) (#16727)
This reverts commit be6aa2e53a17df41abdeb3fa87d86cd740a06e24.

This commit, coupled with the newStyleSynthesis flag (enabled by default on the
v2 branch) do not play nicely together. This is due to the default resources
present in every template (Parameter and Rule for checking bootstrap version).
The `templateMatches` usage introduced in the original PR does not filter out
these default inclusions, leading to most every of those tests to fail. Rather
than block the forward merge on trying to come up with a clever solution,
temporarily reverting.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*