docs(CONTRIBUTING): document guidance for preview APIs
With the commitment to not support breaking changes in stable modules (and all
of aws-cdk-lib), we need to document the standard for introducing APIs we know
to be potentially experimental or unstable.

This standard was originally proposed in https://github.com/aws/aws-cdk-rfcs/blob/master/text/0249-v2-experiments.md, and
was first actually used in https://github.com/aws/aws-cdk/pull/13780.

This task documents the standard in the CONTRIBUTING guide so we can use it as a
reference going forward.

closes #16434