chore(cli): add more capabilities to the hotswap CFN evaluate sub-system (#16696)
The current functionality we use for evaluating CloudFormation in the hotswap part of the CLI is very limited:
only allows substituting the values of parameters used for Assets.
That's not good enough when doing substitutions for StepFunctions State Machines from
[this PR](https://github.com/aws/aws-cdk/pull/16489), for example.

Enhance the capabilities of the CFN eval sub-system by introducing a new class,
`CloudFormationExecutableTemplate`, that allows resolving references to resources inside the template.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*