fix(s3): setting `autoDeleteObjects` to `false` empties the bucket
This was caused by the Custom Resource--which had previously been
deployed when `autoDeleteObjects: true`--being removed when
`autoDeleteObjects` is flipped off again. The custom resource would
indiscriminately empty the bucket as it was being deleted.

Fix by having the custom resource inspect the ongoing CloudFormation
deployment: if the bucket would not be deleted as part of the ongoing
deployment, also do not empty it.

Fixes #16603.