chore: reset jsii-rosetta worker count to default (#16755)
This value was reduced as part of troubleshooting of various Node Worker memory
issues. These issues are theorized to have been mitigated by #16752. Our pack
time is currently over 2 hours, compared to 20-30 minutes prior to the set of
changes. By removing this worker count override, we should be able to get back
to normal pack times and speed up the pipeline.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*