fix(cli): progress bar overshoots count by 1 for stack updates (#16168)
Currently, the `resourcesTotal` output is one short as it doesn't account for the `UPDATE_COMPLETE` event emitted when updating a stack. This PR increases the `resourcesTotal` variable depending on whether the stack is being updated or created.

Noticed this bug when using the CDK on private projects.

This has had a minor fix previously to address the `CREATE_COMPLETE` event emitted when creating a stack, however this did not address the `UPDATE_COMPLETE` event emitted when updating a stack. This caused updated events to produce the following output:

![image](https://user-images.githubusercontent.com/57939433/130373537-5dfacd3c-df7d-4272-abac-a4cf7c04cc47.png)

To address this issue, I:
- Added `+1` to the `resourcesTotal` prop in `packages/aws-cdk/lib/api/deploy-stack.ts` for the `StackActivityMonitor` class depending on whether the stack being deployed already exists using the `cloudFormationStack.exists` boolean.

I also addressed a spacing issue between the pipe (`|`) and the timestamp, as seen in the image above.

Collaborators:
- @JWK95: Provided code review & valid suggestions

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*