chore(individual-pkg-gen): fix bug in setting alpha package visibility (#16787)
The logic to remove the "private" marker for the alpha module `package.json`s
was backported to `master` without any change or an appropriate qualifier. This
leads to the alpha packages being set to public on `master`, which is not what
we want. Rather than introduce v1- and v2-specific logic here, I opted to look
at the current package's setting, and swap it. The logic is that if we're
publishing `aws-foobar`, we don't want to publish `aws-foobar-alpha`, and vice
versa.

Also fixed a bug where alpha'ed packages were being re-alpha'ed when transform
was run multiple times in local development.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*