docs(s3): correct heading levels Object Ownership / Bucket deletion (#16790)
In the README file, the chapters `Object Ownership` and `Bucket deletion` are created as subchapter of `The URL for objects`. In my opinion they do not have a relationship to `The URL for objects`. Probably this hierarchy was created by mistake.

I would suggest to move chapters `Object Ownership` and `Bucket deletion` to the same hierarchy level than `The URL for objects`.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*