feat(assertions): matcher support for `templateMatches()` API (#16789)
The `templateMatches()` API behaved differently from the rest of the
`hasXxx()` and `findXxx()` APIs in that it did not accept a Matcher.
This functionality is generally useful to perform partial matching on
the full template.

Further, users can get confused and assume that the `templateMatches()`
API do support Matchers, as this is the only one that is an exception.

Align this API with the rest of the module's behaviour.

A nice side effect of this is that this module no longer needs to vendor
in changes from the 'assert' module and brings this in line with the
other modules in this repo.

nozem can work again! 🙌

BREAKING CHANGE: The `templateMatches()` API previously performed
an exact match. The default behavior has been updated to be
"object-like".

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*