chore(assertions): replace `absentProperty()` with `absent()` and support it as a `Matcher` type (#16653)
Currently `Match.absentProperty()` is not of type `Matcher` and this introduces a few flaws. For example, `Matcher.isMatcher(Match.absentProperty())` is `false`.

This PR fixes this issue by replacing `Match.absentProperty()` with `Match.absent()` that is of type `Matcher`. The decision to change the name is due to supporting this as a general matcher and not just for Properties.

BREAKING CHANGE: `Match.absentProperty()` becomes `Match.absent()`, and its type changes from `string` to `Matcher`.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*