fix(lambda): currentVersion fails when architecture specified (#16849)
The cause for this is that the new 'Architecture' property added to the
CloudFormation specification is not classified as version locked or not.

Added a test to ensure that any missing properties are caught in the
future.

Further, deprecated the `architectures` property and replaced with a
singular `architecture` prop. Lambda Functions only support one
architecture.

Additionally, removed the CFN spec patch, now that the CloudFormation
resource specification includes the Architecture fields.

fixes #16814


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*