chore(assertions): snippets in readme are now rosetta compilable (#16801)
Rosetta automatic snippet translation is doing a poor
job of snippet translation. This is mainly because the
snippets are not compilable.

Update snippets to make them compilable and translate
correctly.

Additionally, make the rosetta compilation strict and run
during the build step.

There are still a few more translation errors, but these are
arising from bugs in rosetta.
https://github.com/aws/jsii/issues/3026
https://github.com/aws/jsii/issues/3029

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*