fix(aws-servicecatalog): Allow users to create multiple product versions from assets
This small PR should fix an issue brought up that we used a static `Template` as the resource
name for assets which causes a collision if you have multiple versions.  The correct configuration
should be that the asset name is unique for each unique template file uploaded for a product version.

Fixes: https://github.com/aws/aws-cdk/issues/16892