chore(assertions): consistent naming in maven
The maven artifact id is configured to 'cdk-assertions'.

This makes the assertions module naming different from the other package
managers, and hence harder to discover.

Change the artifact id to 'assertions' and make this consistent.

BREAKING CHANGE: Starting this release, the `assertions` module will be
published to Maven with the name 'assertions' instead of
'cdk-assertions'.