fix(aws-servicecatalog): Allow users to create multiple product versions from assets.  (#16914)
This small PR should fix an issue brought up that we used a static `Template` as the resource
name for assets which causes a collision if you have multiple versions.  The correct configuration
should be that the asset name is unique for each unique template file uploaded for a product version.

Fixes: https://github.com/aws/aws-cdk/issues/16892


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*