chore(cdk-release): create our release notes as part of the build
Currently, our bump process creates `CHANGELOG.md` for the `master` branch, and
`CHANGELOG.v2.md` and `CHANGELOG.v2.alpha.md` on the `v2-main` branch. For the
latter, we need to combine the two changelogs into one for the GitHub release,
where all of the artifacts of the release (both aws-cdk-lib and all the alpha
modules) are bundled together.

As proposed in the [RFC], this combined format shows the stable/aws-cdk-lib
changes first, and then the alpha module changes under their own heading.  This
generated -- but not saved -- RELEASE_NOTES.md will be used by our publishing
pipeline to generate the GitHub release notes.

[RFC]: https://github.com/aws/aws-cdk-rfcs/blob/master/text/0249-v2-experiments.md#changelog--release-notes

fixes #16802