chore(iam): reorganize imports to fix circular dependency issue (#16932)
To recreate this issue, create a blank CDK App and add a single Lambda Function to it. Synthesis will fail with the message "Class extends value undefined is not a constructor or null". This is caused by `AnyPrincipal` being imported from `'./'` and other classes from the same file as `AnyPrincipal` (`principals.ts`) being imported from `'./principals'`.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*