feat(apigatewayv2-integrations): http api - support for request parameter mapping (#15630)
----
This is an initial PR as discussed with @nija-at in an attempt to describe the user experience for supporting parameter mapping.
This PR will only support parameter mappings for HTTP APIs _without_ an integration subtypes, but it will provide interfaces that can (and probably should) be reused when adding support for integration subtypes as well. Since it also provides the possibility to provide custom key/value pairs for maximum flexibility, it can support and integration subtype although it requires a bit more work on the user side.

closes #15293 

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*