fix(lambda): docker image function fails when insightsVersion is specified (#16781)
Lambda insights is configured as a set of layers that are then
auto-added to the lambda function when the `insightsVersion` property is
specified.

However, Lambda functions deployed as container images cannot contain
layers. Instead, the user is expected to bring a container image with
the Insights agent pre-installed.

Update the CDK code, so that the layer is not added for Lambda functions
that use container images.

fixes #16642


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*