chore(individual-pkg-gen): set the alpha version in alpha module's package.json files if present (#16965)
This change was already approved and merged in: https://github.com/aws/aws-cdk/pull/16322. It somehow got removed from the `v2-main` branch since originally being merged. This PR is just a cherry-pick of the original commit. 

---

This change sets the `version` key in each alpha module's `package.json` file to the alphaVersion that was created in this PR: https://github.com/aws/aws-cdk/pull/16043/

And, also sets the version of each dependency on another alpha module to the same version.

Depends on: #16321

Part of: #15591

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*