chore(assertions): remove `rosetta:extract` from build command
This leads to the generation of `.jsii.tabl.json` during build,
which (by lack of an NPM include list and it being missing from the
NPM exclude list) gets subsequently published to NPM...

...and finally interferes with the jsii pipeline where we build
the latest version of CDK, and the presence of a file generated
with an old version of `rosetta` cannot be read by a new version
of `rosetta`.

There will be a supported mechanism to achieve what this
single post-build command is trying to achieve.